Create README.md
early stages, so far have been limited to collecting a few parts such as an esp32 board, an Arduino mega 2650 board an Arduino CNC shield, a kit of 5 servos with servo control boards, and a kit containing 3 stepper motors with some mounting hardware 8mm rods with linear bearings and some miscellaneous other 'stuff'. i have downloaded the Arduino IDE and a few example programs.  made the steppers move in a predictable way but have not done any of my own programming yet.  I'm looking for a program to convert .svg files to a gcode that the Arduino can interpret.